\name{eqtlversion}
\alias{eqtlversion}

\title{Installed version of R/eqtl}

\description{
Print the version number of the currently intalled version of R/eqtl
}

\usage{eqtlversion()}

\value{
A character string with the version number of the currently installed version of R/eqtl.
}

\author{Hamid A Khalili, \email{hamid.khalili@gmail.com} }

\seealso{ \code{\link[base]{version}} }

\keyword{print}
