\name{seed10}
\alias{seed10}
\docType{data}
\title{Data on gene expression level variation}
\description{Data from an experiment on the expression level variation in the \emph{A.thaliana} seed at 10 day after pollination.}
\usage{data(seed10)}
\format{An object of class \code{cross} and \code{riself}.}
\details{
There are 420 RIL individuals typed at 69 markers and 160 individuals have been retained for phenotyping. The population is the 33RV Versailles RIL population Bay x Sha. See references below. The phenotype is a sample of 500 Gene Sequence Tags hybridization signals measured on the CATMA microarray. See references below.
}
\source{
Jean-Pierre Renou and Alain Lecharny (CATdb a Complete Arabidopsis Transcriptome database)
\url{http://urgv.evry.inra.fr/CATdb}\cr
Loudet (Genetic and Plant breeding, the VAST lab, INRA VERSAILLES)\cr
\url{http://dbsgap.versailles.inra.fr/vnat/Documentation/33/DOC.html}
}
\references{
Crowe M.L. et al., (2003) CATMA,  A complete Arabidopsis GST database. Nucleic Acids Res 31:156-158.\cr
Loudet et al.(2002) Theoretical and Applied Genetics, vol 104, pp 1173-1184
}
\examples{data(seed10)}
\keyword{datasets}
